Fix: Moved keycloak provider into @providers.tsx
